renaming, loading multiple models and ability to save audio
