+++ src/python/assistant/apps/torchserve/torchserve.config
@@ -1,7 +1,7 @@
 inference_address=http://0.0.0.0:8080
 default_workers_per_model=1
 
-load_models=jenkins_prompt.mar
+load_models=jenkins_prompt.mar, jenkins_listen.mar
 models = {\
  jenkins_prompt : {\
     "1.0": {\
